Add logic for default images folder.
This required several steps. A default folder parm was added to runtime parms.
The image folder defaults to that (it can include % formats which need to be
parsed). During entry (create or edit) a blank image value is set to the
default. The value of the URL is not tested during validation if it equals the default.
Also, the default values is cleared before writing to the DB.

Validation routine was enhanced to use a single parameter for data type
(text, date, graphic, or numeric).